Move colorPaletteTypes to schemas/
