+++ src/constants/coreSettings.ts
@@ -1,7 +1,7 @@
 import { LinkMarkerShape } from '@comfyorg/litegraph'
 import { LiteGraph } from '@comfyorg/litegraph'
 
-import type { ColorPalettes } from '@/types/colorPaletteTypes'
+import type { ColorPalettes } from '@/schemas/colorPaletteTypes'
 import type { Keybinding } from '@/types/keyBindingTypes'
 import { NodeBadgeMode } from '@/types/nodeSource'
 import { LinkReleaseTriggerAction } from '@/types/searchBoxTypes'
