+++ src/schemas/apiTypes.ts
@@ -2,8 +2,8 @@ import { LinkMarkerShape } from '@comfyorg/litegraph'
 import { ZodType, z } from 'zod'
 import { fromZodError } from 'zod-validation-error'
 
+import { colorPalettesSchema } from '@/schemas/colorPaletteTypes'
 import { zComfyWorkflow, zNodeId } from '@/schemas/comfyWorkflow'
-import { colorPalettesSchema } from '@/types/colorPaletteTypes'
 import { zKeybinding } from '@/types/keyBindingTypes'
 import { NodeBadgeMode } from '@/types/nodeSource'
 import { LinkReleaseTriggerAction } from '@/types/searchBoxTypes'
