+++ src/schemas/colorPaletteTypes.ts
